+++ boop/layereddict.py
@@ -1,0 +1,1 @@
+class LayeredDict(object):